+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Onewire plugin, reads onewire sensor data from /sys/bus/w1/devices/
+
 ## [0.4.0] - 2015-11-26
 ### Added
 - pellmonconf, a web based text editor for the configuration files